+++ korona/html/tags/acronym.py
@@ -16,7 +16,7 @@ class Acronym(object):
     .. versionadded:: 0.1.0
 
     .. versionchanged:: 0.2.0
-        Renamed :func:`construct_tag()` to :func:`construct()`.
+        Renamed :func:`construct_tag` to :func:`construct`.
     """
     def __init__(self, text=None):
         self.tag = 'acronym'